move feature components to new folder
