+++ src/components/feature_components/PersonalAssessmentFeature.js
@@ -1,11 +1,8 @@
 import React from "react";
 import { makeStyles } from "@material-ui/core/styles";
 import Card from "@material-ui/core/Card";
-import CardActionArea from "@material-ui/core/CardActionArea";
-import CardActions from "@material-ui/core/CardActions";
 import CardContent from "@material-ui/core/CardContent";
-import CardMedia from "@material-ui/core/CardMedia";
-import assessmentImg from "../assets/assessment.jpg";
+import assessmentImg from "../../assets/assessment.jpg";
 import Typography from "@material-ui/core/Typography";
 
 const useStyles = makeStyles({
